requirements.txt
discord.py
google==3.0.0
psycopg2-binary==2.8.6
python-dotenv==0.15.0


